Update
Did an update to the last version of Eclipse so it made some projects
(literally every one of them) incompatible with the new version. Updated
all the projects to the new versions and added a new project which is
Guess The Number (GTN), Interactive Keyboard (IK) and Second Window (SW)
but all in one project, apart from those projects separate. Nothing
changed for Calculator Interface.